dev: modularize styles
