+++ packages/admin/components/notices/styles.module.scss
@@ -1,5 +1,5 @@
 .notices {
-	:global( .components-snackbar ) {
+	:global(.components-snackbar) {
 		bottom: 0.5rem;
 		position: fixed;
 	}
