+++ packages/admin/layout/header/menu/styles.module.scss
@@ -4,14 +4,14 @@
 
 .linkIcon {
 	transform: rotate(-45deg);
-	opacity: .85;
+	opacity: 0.85;
 }
 
 :global(.components-button) {
 	&.active {
 		border-bottom: 2px solid;
 		border-image: linear-gradient(to right, transparent 0%, var(--wp-admin-theme-color-darker-20) 100%) 1;
-		
+
 		&:global(.is-tertiary) {
 			color: var(--wp-admin-theme-color-darker-20);
 		}
